[Feat] Adding Inventory  , TransactionMiddleware , UnitOfWork , InventoryValidator

+++ InventoryManagementSystem.Core/DTOs/ProductDto/GetProductResponse.cs
@@ -8,7 +8,7 @@ namespace InventoryManagementSystem.Core.DTOs.ProductDto
 {
     public class GetProductResponse
     {
-        public int ProductId { get; set; }
+        public decimal ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
         public decimal ProductPrice { get; set; }
